Use 16 bit integers as appropriate.
Instead of using many 32 bit integers to hold 16 bit values,
use uint16_t everywhere that's holding the raw data.  Do not
munge into signed values early.  Move baseline subtraction
out to just before we write out the results, while still using
the baseline substracted value for thresholds.  This is a little
more work, but it is safer and more clear.

+++ include/USBstream-TypeDef.h
@@ -1,1 +1,1 @@
-typedef std::vector<std::vector<int32_t> > DataVector;
+typedef std::vector<std::vector<uint16_t> > DataVector;
